refactor(levelup): rename controller_test
rename controller_test -> controller in /main/test.. package
