+++ src/test/java/com/greenfoxacademy/levelup/controller/PitchesRestControllerTest.java
@@ -1,4 +1,4 @@
-package com.greenfoxacademy.levelup.controller_test;
+package com.greenfoxacademy.levelup.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -8,7 +8,6 @@ import com.greenfoxacademy.levelup.collection.Message;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
